Replace test data

+++ previewer/main.js
@@ -8,8 +8,8 @@ const readdir = util.promisify(fs.readdir)
 const writeFile = util.promisify(fs.writeFile)
 const readFile = util.promisify(fs.readFile)
 
-const inputPath = process.env.INPUT_DATA_PATH || "./input"
-const outputPath = process.env.OUTPUT_DATA_PATH || "./output"
+const inputPath = process.env.INPUT_DATA_PATH || "../test-data/input"
+const outputPath = process.env.OUTPUT_DATA_PATH || "../test-data/output"
 
 function sleep(t) {
   return new Promise((resolve) => setTimeout(resolve, t))
@@ -26,30 +26,34 @@ if (!fs.existsSync(`${outputPath}/data`)){
 const processTrip = async tripId => {
   console.log("Process trip", tripId)
 
-  const browser = await puppeteer.launch({
-    defaultViewport: {width: 800, height: 600}
-  });
-  const page = await browser.newPage();
-  page.on('console', msg => console.log('CONSOLE LOG:', msg.text()));
-  page.on('requestfailed', msg => console.log('REQUEST FAILED LOG:', msg.text()));
-  
-  await page.goto(`http://localhost:8080/#/maps/${tripId}`);
-  await sleep(10000)
-  await page.screenshot({path: `${outputPath}/data/${tripId}/preview.png`});
+  const out = `${outputPath}/data/${tripId}/preview.png`
+  // If preview doesn't exist we create a new one.
+  if (!fs.existsSync(out)){
+    const browser = await puppeteer.launch({
+      defaultViewport: {width: 800, height: 600}
+    });
+    const page = await browser.newPage();
+    page.on('console', msg => console.log('CONSOLE LOG:', msg.text()));
+    page.on('requestfailed', msg => console.log('REQUEST FAILED LOG:', msg.text()));
+    
+    await page.goto(`http://localhost:8080/#/maps/${tripId}`);
+    await sleep(5000)
+    await page.screenshot({path: out});
 
-  await browser.close()
+    await browser.close()
 
-  const tripInfo = yaml.safeLoad(fs.readFileSync(`${inputPath}/${tripId}/trip.yml`, 'utf8'))
+    const tripInfo = yaml.safeLoad(fs.readFileSync(`${inputPath}/${tripId}/trip.yml`, 'utf8'))
 
-  const output = mustache.to_html(
-    await readFile(`template.mustache`, 'utf8'),
-    {
-      title: tripInfo.title,
-      description: tripInfo.description,
-      tripId: tripId
-    }
-  )
-  await writeFile(`${outputPath}/trips/${tripId}.html`, output, 'utf8')
+    const output = mustache.to_html(
+      await readFile(`template.mustache`, 'utf8'),
+      {
+        title: tripInfo.title,
+        description: tripInfo.description,
+        tripId: tripId
+      }
+    )
+    await writeFile(`${outputPath}/trips/${tripId}.html`, output, 'utf8')
+  }
 }
 
 // Read list of trips.
